fix for qa breaking, re-apply onboard patch
