+++ package.json
@@ -199,12 +199,13 @@
         "@react-native-async-storage/async-storage": "1.17.10",
         "@lingui/react": "3.15.0",
         "@lingui/core": "3.15.0",
-        "@web3-onboard/core@^2.12.1": "patch:@web3-onboard/core@npm%3A2.12.1#./.yarn/patches/@web3-onboard-core-npm-2.12.1-94d32823fe.patch"
+        "@web3-onboard/core@^2.12.1": "patch:@web3-onboard/core@npm%3A2.12.1#./.yarn/patches/@web3-onboard-core-npm-2.12.1-94d32823fe.patch",
+        "@web3-onboard/core@^2.13.0": "patch:@web3-onboard/core@npm%3A2.13.0#./.yarn/patches/@web3-onboard-core-npm-2.13.0-74decc51bc.patch"
     },
     "dependencies": {
         "@gooddollar/good-design": "^0.0.27",
         "@gooddollar/goodprotocol": "1.0.30-beta.3",
-        "@gooddollar/web3sdk": "^0.1.15",
+        "@gooddollar/web3sdk": "^0.1.17",
         "@gooddollar/web3sdk-v2": "^0.1.59",
         "@headlessui/react": "1.5.0",
         "@types/react-text-mask": "^5.4.8",
